Rename the file to avoid confusion that it contains a store
