+++ src/lib/entities/episode/index.ts
@@ -1,3 +1,3 @@
 export { default as EpisodeCardShell } from './ui/episode-card-shell.svelte';
 export { currentlyPlayingEpisode } from './model/currently-playing-store';
-export { getTranscript } from './model/transcript-store';
+export { getTranscript } from './model/transcripts';
